Rename "download.py" to "update.py"
